Update README.md
Original Text:
All UI code is shifted onto fragments, except the welcome screen.
The game logic object utilizes singleton model.
Revision:
The code architecture of note are as follows:
1.	All UI code is shifted onto fragments from activities, with the exception of the welcome screen.
2.	The game logic object utilizes singleton model that restricts the instantiation of a class to one object, so that there is no race condition with multiple instances of the game running.
Rationale:
The code section is given a header for better coherence. Additional explanation of the code architecture is provided for the audience which might not be familiar why the UI code is shifted onto fragments. The purpose of the singleton model is explained.